pyproject.toml
[tool.poetry]
name = "dream-interpreter"
version = "0.1.0"
description = "AI-powered dream interpretation with symbolic visualisation"
authors = ["Charlotte Rogers <132924201+Chazzy11@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
pillow = "^10.0.0"
textblob = "^0.17.1"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
httpx = "^0.25.0"

[tool.black]
line-length = 88
target-version = ['py39']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"